requirements.txt
pandas
bs4
requests
selenium
requests~=2.31.0
Flask~=2.3.2
Werkzeug~=2.3.6
Jinja2~=3.1.2
WTForms~=3.0.1
gunicorn==21.2.0
Flask_Login==0.6.2
flask_sqlalchemy==3.0.5
Bootstrap_Flask==2.2.0
Flask_WTF==1.1.1
langchain-groq
python-dotenv
langchain
faiss-cpu
markdown